requirements.txt
torch == 2.3.1
pillow == 10.4.0
torchvision == 0.19.0